Dev pool

+++ backend/routes/bookings.py
@@ -1,3 +1,4 @@
+from datetime import datetime, timedelta
 from uuid import UUID
 from fastapi import APIRouter, Depends, HTTPException, status
 from sqlalchemy.orm import Session
@@ -29,6 +30,16 @@ def create_booking(
     if existing:
         raise HTTPException(status_code=400, detail="Booking already exists")
 
+    class_obj = db.query(class_.Class).get(booking_data.class_id)
+    if not class_obj:
+        raise HTTPException(status_code=404, detail="Class not found")
+
+    class_datetime_str = f"{class_obj.date} {class_obj.time}"
+    class_datetime = datetime.strptime(class_datetime_str, "%Y-%m-%d %H:%M")
+
+    if class_datetime - datetime.now() < timedelta(hours=1.5):
+        raise HTTPException(status_code=400, detail="Booking cannot be booked")
+
     # creates new booking
     new_booking = booking.Booking(
         user_id = user_id,
@@ -40,10 +51,8 @@ def create_booking(
     db.commit()
     db.refresh(new_booking)
 
-    # joined class_
-    db.refresh(new_booking)
-    db.expunge(new_booking)
-    new_booking.class_ = db.query(class_.Class).get(new_booking.class_id)
+    new_booking.class_ = class_obj
+    class_obj.current_participants = len([b for b in class_obj.bookings if b.user])
 
     return new_booking
 
@@ -58,7 +67,17 @@ def cancel_booking(
         raise HTTPException(status_code=404, detail="Booking not found")
     
     if booking_obj.user_id != current_user.id:
-        raise HTTPException(403, "Not allowed to cancel this booking")
+        raise HTTPException(status_code=403, detail="Not allowed to cancel this booking")
+
+    class_obj = db.query(class_.Class).get(booking_obj.class_id)
+    if not class_obj:
+        raise HTTPException(status_code=404, detail="Class not found")
+    
+    class_datetime_str = f"{class_obj.date} {class_obj.time}"
+    class_datetime = datetime.strptime(class_datetime_str, "%Y-%m-%d %H:%M")
+
+    if class_datetime - datetime.now() < timedelta(hours=2):
+        raise HTTPException(status_code=400, detail="Booking cannot be canceled")
 
     db.delete(booking_obj)
     db.commit()
